fastapi-engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
sqlalchemy[asyncio]>=2.0.31
asyncpg>=0.29.0
alembic>=1.13.2
redis[hiredis]>=5.0.7
celery>=5.4.0
neo4j>=5.22.0
langchain>=0.2.6
langchain-google-genai>=1.0.0
langchain-xai>=0.1.0
langchain-community>=0.2.6
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
tree-sitter>=0.22.3
google-generativeai>=0.5.0
python-dotenv>=1.0.1
GitPython>=3.1.43
